requirements.txt
djangorestframework==3.14.4
dj-database-url==0.5.0
Django==3.2.4
gunicorn==20.1.0
psycopg2-binary==2.9.1
whitenoise==2.9.1

django-cors-headers==3.7.0
dj-rest-auth==1.1.0
django-allauth==0.42.0
